docとテスト修正
